+++ src/components/workout-plans/forms/WorkoutSessionForm.tsx
@@ -6,7 +6,7 @@ import {
 	TextField,
 	Typography
 } from '@mui/material';
-import { FC, useState } from 'react';
+import { FC, useEffect, useState } from 'react';
 import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
 import { Delete } from '@mui/icons-material';
 
@@ -21,49 +21,6 @@ type WorkoutSessionFormProps = {
 	onUpdate: (updatedWorkoutSession?: WorkoutSession) => void;
 };
 
-const validateExerciseVolume = (exerciseVolume: ExerciseVolume) => {
-	if (exerciseVolume.usesRange) {
-		return exerciseVolume.values.min < exerciseVolume.values.max;
-	}
-	return true;
-};
-
-export const validateExercise = (exercise: Exercise) => {
-	if (!validateExerciseVolume(exercise.reps)) {
-		return false;
-	}
-
-	if (!validateExerciseVolume(exercise.sets)) {
-		return false;
-	}
-
-	return exercise.name.length !== 0;
-};
-
-const validateWorkoutSession = (workout: WorkoutSession) => {
-	if (workout.exercises.length === 0) {
-		alert(
-			`Workout session ${
-				workout.id + 1
-			} does not contain any exercises, please add at least one exercise, or delete the session`
-		);
-		return false;
-	}
-
-	for (let i = 0; i < workout.exercises.length; ++i) {
-		if (!validateExercise(workout.exercises[i])) {
-			alert(
-				`Exercise ${i + 1} in workout session ${
-					workout.id + 1
-				} contains invalid values`
-			);
-			return false;
-		}
-	}
-
-	return true;
-};
-
 const WorkoutSessionForm: FC<WorkoutSessionFormProps> = ({
 	elemIndex,
 	workoutSession,
@@ -75,8 +32,9 @@ const WorkoutSessionForm: FC<WorkoutSessionFormProps> = ({
 	);
 	const [exercises, setExercises] = useState<Exercise[]>([]);
 
-	const [_, setWorkoutSessionObj] = useState<WorkoutSession>(workoutSession);
-	const [locked, setLocked] = useState<boolean>(false);
+	const [_, setWorkoutSessionObj] = useState<WorkoutSession>(workoutSession); // todo check if this is even useful
+
+	// const [locked, setLocked] = useState<boolean>(false);
 
 	const renderExercises = () => {
 		if (exercises.length === 0) {
@@ -88,7 +46,7 @@ const WorkoutSessionForm: FC<WorkoutSessionFormProps> = ({
 				key={`${elemIndex}-exercise-${i}`}
 				elemIndex={`${elemIndex}-exercise-${i}`}
 				exercise={exercise}
-				isLocked={locked}
+				// isLocked={locked}
 				onUpdate={updatedExercise => onUpdateExercise(i, updatedExercise)}
 			/>
 		));
@@ -128,19 +86,18 @@ const WorkoutSessionForm: FC<WorkoutSessionFormProps> = ({
 		setExercises([...exercises, newExercise]);
 	};
 
+	useEffect(() => {
+		updateWorkoutSession(); //todo check
+	}, [workoutSessionName, exercises]);
+
 	const updateWorkoutSession = () => {
-		setLocked(true);
+		// setLocked(true);
 
 		const updatedWorkoutSession: WorkoutSession = {
 			id: workoutSession.id,
 			name: workoutSessionName,
 			exercises
 		};
-
-		if (!validateWorkoutSession(updatedWorkoutSession)) {
-			setLocked(false);
-			return;
-		}
 
 		setWorkoutSessionObj(updatedWorkoutSession);
 		onUpdate(updatedWorkoutSession);
@@ -157,16 +114,19 @@ const WorkoutSessionForm: FC<WorkoutSessionFormProps> = ({
 				placeholder="(Optional) Enter a custom workout session name"
 				{...workoutSessionNameProps}
 				type={`${elemIndex}-workout-session-name`}
-				disabled={locked}
+				// disabled={locked}
 				sx={{ marginBottom: '0.2rem' }}
 			/>
 			{renderExercises()}
-			<IconButton onClick={addExercise} disabled={locked}>
+			<IconButton
+				onClick={addExercise}
+				// disabled={locked}
+			>
 				<AddCircleOutlineIcon />
 				<Typography>Add exercise</Typography>
 			</IconButton>
 			<Box sx={{ display: 'flex', justifyContent: 'space-between' }}>
-				{locked ? (
+				{/* {locked ? (
 					<Button
 						onClick={() => {
 							setLocked(false);
@@ -182,7 +142,7 @@ const WorkoutSessionForm: FC<WorkoutSessionFormProps> = ({
 					>
 						Confirm workout
 					</Button>
-				)}
+				)} */}
 				<IconButton onClick={removeWorkoutSession}>
 					<Delete />
 					<Typography>Delete workout</Typography>
